+++ packages/build-utils/test/unit.test.ts
@@ -818,6 +818,17 @@ it('should detect turborepo project supporting corepack', async () => {
   expect(result.turboSupportsCorepackHome).toEqual(true);
 });
 
+it('should handle turborepo project with comments in turbo.json', async () => {
+  const base = path.join(
+    __dirname,
+    'fixtures',
+    '43-turborepo-with-comments-in-turbo-json'
+  );
+  const fixture = path.join(base, '/apps/web');
+  const result = await scanParentDirs(fixture, true, base);
+  expect(result.turboSupportsCorepackHome).toEqual(true);
+});
+
 it('should detect turborepo project not supporting corepack', async () => {
   const base = path.join(
     __dirname,
@@ -834,6 +845,41 @@ it('should detect non-turborepo monorepo', async () => {
   const fixture = path.join(base, '/c');
   const result = await scanParentDirs(fixture, true, base);
   expect(result.turboSupportsCorepackHome).toEqual(undefined);
+});
+
+it('should detect `packageManager` in npm monorepo', async () => {
+  try {
+    process.env.ENABLE_EXPERIMENTAL_COREPACK = '1';
+
+    const base = path.join(__dirname, 'fixtures', '41-npm-workspaces-corepack');
+    const fixture = path.join(base, 'a');
+    const result = await scanParentDirs(fixture, false, base);
+    expect(result.cliType).toEqual('npm');
+    expect(result.packageJsonPackageManager).toEqual('npm@10.7.0');
+    expect(result.lockfileVersion).toEqual(undefined);
+    expect(result.packageJsonPath).toEqual(path.join(fixture, 'package.json'));
+  } finally {
+    delete process.env.ENABLE_EXPERIMENTAL_COREPACK;
+  }
+});
+
+it('should detect `packageManager` in pnpm monorepo', async () => {
+  try {
+    process.env.ENABLE_EXPERIMENTAL_COREPACK = '1';
+    const base = path.join(
+      __dirname,
+      'fixtures',
+      '42-pnpm-workspaces-corepack'
+    );
+    const fixture = path.join(base, 'c');
+    const result = await scanParentDirs(fixture, false, base);
+    expect(result.cliType).toEqual('pnpm');
+    expect(result.packageJsonPackageManager).toEqual('pnpm@8.3.1');
+    expect(result.lockfileVersion).toEqual(undefined);
+    expect(result.packageJsonPath).toEqual(path.join(fixture, 'package.json'));
+  } finally {
+    delete process.env.ENABLE_EXPERIMENTAL_COREPACK;
+  }
 });
 
 it('should retry npm install when peer deps invalid and npm@8 on node@16', async () => {