+++ packages/remix/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/remix-builder
 
+## 2.2.12
+
+### Patch Changes
+
+- Fix corepack `packageManager` detection on monorepos ([#12258](https://github.com/vercel/vercel/pull/12258))
+
 ## 2.2.11
 
 ### Patch Changes